feat: game state reducer and component
