+++ client/src/helpers/constants.ts
@@ -1,0 +1,8 @@
+export const DIE_VALUE_DICT = {
+  1: 1,
+  2: 2,
+  0: 3,
+  4: 4,
+  5: 5,
+  6: 6,
+};
